+++ Enums/CommonID.cs
@@ -2,13 +2,13 @@
 
 public enum CommonID
 {
-    CPU_Temperature_CPU_Max,
-    CPU_Multiplier_CPU_Max,
-    CPU_Core_Clock_CPU_Max,
-    CPU_Temperature_CPU_Average,
-    CPU_Multiplier_CPU_Average,
-    CPU_Core_Clock_CPU_Average,
-    CPU_Temperature_CPU_Min,
-    CPU_Multiplier_CPU_Min,
-    CPU_Core_Clock_CPU_Min,
+    CPU_Temperature_Temperature_Max,
+    CPU_Multiplier_Multiplier_Max,
+    CPU_Frequency_Core_Clock_Max,
+    CPU_Temperature_Temperature_Average,
+    CPU_Multiplier_Multiplier_Average,
+    CPU_Frequency_Core_Clock_Average,
+    CPU_Temperature_Temperature_Min,
+    CPU_Multiplier_Multiplier_Min,
+    CPU_Frequency_Core_Clock_Min,
 }