+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.1.7 - 2025-05-19
+
+### Changed
+
+- Optimized communication between Plugin and ArgusMonitorLink
+- Reduced Ram usage
+- Optimized ArgusMonitorLink
+- Optimized the compiler
+- Adjusted Argus Data API poll rate to be in sync with the plugin to reduce cpu load
+- Added poll delay setting
+
 ## 0.1.6 - 2025-05-16
 
 ### Changed